ember-data-strepla: Extract `BaseSerializer` class
