+++ lib/ember-data-strepla/app/serializers/strepla-competition-class.js
@@ -1,10 +1,8 @@
-import { getOwner } from '@ember/application';
 import { assert } from '@ember/debug';
-import DS from 'ember-data';
 
-const { Serializer } = DS;
+import BaseSerializer from './-base';
 
-export default class extends Serializer {
+export default class extends BaseSerializer {
   normalizeResponse(store, primaryModelClass, payload, id, requestType) {
     assert(`Unknown \`requestType\`: ${requestType}`, ['query'].indexOf(requestType) !== -1);
 
@@ -32,27 +30,5 @@ export default class extends Serializer {
     this.applyTransforms(modelClass, data.attributes);
 
     return { data, included: [] };
-  }
-
-  applyTransforms(typeClass, data) {
-    let attributes = typeClass.attributes;
-
-    typeClass.eachTransformedAttribute((key, typeClass) => {
-      if (data[key] !== undefined) {
-        let transform = this.transformFor(typeClass);
-        let transformMeta = attributes.get(key);
-        data[key] = transform.deserialize(data[key], transformMeta.options);
-      }
-    });
-
-    return data;
-  }
-
-  transformFor(attributeType) {
-    let transform = getOwner(this).lookup(`transform:${attributeType}`);
-
-    assert(`Unable to find the transform for \`attr('${attributeType}')\``, transform);
-
-    return transform;
   }
 }
